Added an option to support RSME and RSS for L-Method
RMSE was proposed by the original paper, RSS is a simplification for a
speeding up the code.
For most practical cases RSS should have a similar accuracy to RMSE.

This should help with #18.

One important thing to consider, the code in this library does not
consider unit spacing between x points.